refactor: move deleteFolderRecursive to FileFunctions
