+++ extension/src/Documentation.ts
@@ -11,8 +11,8 @@ import {
   DocsType,
 } from "./ALObject/Enums";
 import { ALProcedure } from "./ALObject/ALProcedure";
-import { deleteFolderRecursive, formatDate, replaceAll } from "./Common";
-import { createFolderIfNotExist } from "./FileFunctions";
+import { formatDate, replaceAll } from "./Common";
+import { deleteFolderRecursive, createFolderIfNotExist } from "./FileFunctions";
 import xmldom = require("@xmldom/xmldom");
 import { ALTenantWebService } from "./ALObject/ALTenantWebService";
 import { ALXmlComment } from "./ALObject/ALXmlComment";
