+++ extension/src/FileFunctions.ts
@@ -90,6 +90,23 @@ export function mkDirByPathSync(targetDir: string): string {
   }, initDir);
 }
 
+export function deleteFolderRecursive(directoryPath: string): void {
+  if (fs.existsSync(directoryPath)) {
+    fs.readdirSync(directoryPath).forEach((file) => {
+      const curPath = path.join(directoryPath, file);
+      if (fs.lstatSync(curPath).isDirectory()) {
+        // recurse
+        deleteFolderRecursive(curPath);
+      } else {
+        // delete file
+        fs.unlinkSync(curPath);
+      }
+    });
+
+    fs.rmdirSync(directoryPath);
+  }
+}
+
 export function createFolderIfNotExist(folderPath: string): void {
   if (!fs.existsSync(folderPath)) {
     mkDirByPathSync(folderPath);
