+++ extension/src/test/Common.test.ts
@@ -1,12 +1,7 @@
 import * as assert from "assert";
-import * as path from "path";
-import * as fs from "fs";
 import * as Common from "../Common";
 
 suite("Common", function () {
-  const parentPath = path.resolve(__dirname, "common-test");
-  const newPath = path.resolve(parentPath, "new/path/", Common.formatDate());
-
   test("convertLinefeedToBr", function () {
     assert.strictEqual(
       Common.convertLinefeedToBR("\n\n\n"),
@@ -27,20 +22,6 @@ suite("Common", function () {
       Common.formatDate(d),
       "2001-01-01",
       "Incorrect date string returned"
-    );
-  });
-
-  test("deleteFolderRecursive", function () {
-    Common.deleteFolderRecursive(parentPath);
-    assert.strictEqual(
-      fs.existsSync(newPath),
-      false,
-      "Child path should be deleted"
-    );
-    assert.strictEqual(
-      fs.existsSync(parentPath),
-      false,
-      "Parent path should be deleted."
     );
   });
 });
